fix: eliminate compiler warnings and configure warning suppression
- Fix 30+ warnings in Sidechain source code:
  - Async.cpp: Fix 5 lambda capture shadow warnings
  - UIHelpers.cpp: Fix 8 implicit int-to-float conversion warnings
  - WaveformGenerator.cpp: Fix 4 implicit int-to-float conversions
  - StringFormatter.cpp: Fix 3 int64_t-to-double conversions
  - OSNotification_mac.mm: Fix 3 nullable-to-nonnull conversions
  - Image.cpp: Fix 1 implicit float conversion
  - FeedPost: Add explicit copy constructor/assignment, fix switch-enum

- Configure comprehensive warning suppression in CompilerFlags.cmake:
  - Suppress JUCE-specific warnings that cannot be fixed
  - Add flags for deprecated APIs, size_t conversions, macro braces
  - Add flags for audio processing math, lambda captures, macOS APIs
  - Add flags for generated code safety (missing field initializers)

Plugin now builds cleanly with -Wall -Wextra -Wpedantic enabled.
Only 4 unavoidable warnings remain from JUCE's third-party AU/VST SDK.

+++ plugin/src/ui/profile/SavedPosts.cpp
@@ -199,8 +199,8 @@ void SavedPosts::drawHeader(juce::Graphics& g)
 
     // Bottom border
     g.setColour(Colors::border());
-    g.drawLine(bounds.getX(), bounds.getBottom() - 1,
-               bounds.getRight(), bounds.getBottom() - 1, 1.0f);
+    g.drawLine(static_cast<float>(bounds.getX()), static_cast<float>(bounds.getBottom() - 1),
+               static_cast<float>(bounds.getRight()), static_cast<float>(bounds.getBottom() - 1), 1.0f);
 }
 
 void SavedPosts::drawLoadingState(juce::Graphics& g, juce::Rectangle<int> bounds)
